pyproject.toml
[project]
name = "luna_chat"
version = "0.1.0"
description = "A Creative Assistant."
authors = [
    { name = "Leo Huber", email = "leo_huber@icloud.com" }
]
dependencies = [
    "textual[syntax]==1.0.0",
    "sqlmodel>=0.0.9",
    "humanize>=4.6.0",
    "click>=8.1.6",
    "xdg-base-dirs>=6.0.1",
    "aiosqlite>=0.20.0",
    "click-default-group>=1.2.4",
    "greenlet>=3.0.3",
    "google-generativeai>=0.5.3",
    "pyperclip>=1.8.2",
    "litellm>=1.55.3",
    "pydantic>=2.9.0",
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langchain-text-splitters>=0.3",
    "langchain-community>=0.3",
    "langchain-openai>=0.2",
    "jinja2>=3.1.5",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
luna = "luna_chat.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 150

[tool.uv]
managed = true
dev-dependencies = [
    "mypy>=1.14.1",
    "types-peewee>=3.16.0.0",
    "pre-commit>=4.0.1",
    "textual-dev>=1.7.0",
    "pyinstrument>=5.0.0",
    "ruff>=0.8.6",
    "ipykernel>=6.29.5",
    "pytest>=8.3.5"
]

[tool.mypy]
ignore_missing_imports = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["luna_chat"]
exclude = [
  "/.github",
  "/test",
]
